requirements.txt
numpy==1.24.3
openai-whisper==20231117
deep-translator==1.11.4
edge-tts==7.0.0
moviepy==1.0.3
torch==2.1.2
decorator>=4.0.2
imageio>=2.5
imageio-ffmpeg>=0.4.3
proglog>=0.1.9
tqdm>=4.11.2
PyQt5==5.15.9